+++ src/aks-preview/azext_aks_preview/_params.py
@@ -52,6 +52,9 @@ def load_arguments(self, _):
         c.argument('dns_service_ip')
         c.argument('docker_bridge_address')
         c.argument('load_balancer_sku')
+        c.argument('load_balancer_managed_outbound_ip_count', type=int)
+        c.argument('load_balancer_outbound_ips')
+        c.argument('load_balancer_outbound_ip_prefixes')
         c.argument('enable_addons', options_list=['--enable-addons', '-a'])
         c.argument('disable_rbac', action='store_true')
         c.argument('enable_rbac', action='store_true', options_list=['--enable-rbac', '-r'],
@@ -68,10 +71,12 @@ def load_arguments(self, _):
         c.argument('enable_cluster_autoscaler', action='store_true')
         c.argument('min_count', type=int, validator=validate_nodes_count)
         c.argument('max_count', type=int, validator=validate_nodes_count)
-        c.argument('enable_vmss', action='store_true')
+        c.argument('vm_set_type')
         c.argument('node_zones', zones_type, options_list='--node-zones', help='(PREVIEW) Space-separated list of availability zones where agent nodes will be placed.')
         c.argument('enable_pod_security_policy', action='store_true')
         c.argument('node_resource_group')
+        c.argument('enable_acr')
+        c.argument('acr')
 
     with self.argument_context('aks update') as c:
         c.argument('enable_cluster_autoscaler', options_list=["--enable-cluster-autoscaler", "-e"], action='store_true')
@@ -79,9 +84,15 @@ def load_arguments(self, _):
         c.argument('update_cluster_autoscaler', options_list=["--update-cluster-autoscaler", "-u"], action='store_true')
         c.argument('min_count', type=int, validator=validate_nodes_count)
         c.argument('max_count', type=int, validator=validate_nodes_count)
+        c.argument('load_balancer_managed_outbound_ip_count', type=int)
+        c.argument('load_balancer_outbound_ips')
+        c.argument('load_balancer_outbound_ip_prefixes')
         c.argument('api_server_authorized_ip_ranges', type=str, validator=validate_ip_ranges)
         c.argument('enable_pod_security_policy', action='store_true')
         c.argument('disable_pod_security_policy', action='store_true')
+        c.argument('enable_acr')
+        c.argument('disable_acr')
+        c.argument('acr')
 
     with self.argument_context('aks scale') as c:
         c.argument('nodepool_name', type=str,
@@ -118,6 +129,11 @@ def load_arguments(self, _):
         c.argument('addons', options_list=['--addons', '-a'])
         c.argument('subnet_name', options_list=['--subnet-name', '-s'])
 
+    with self.argument_context('aks get-credentials') as c:
+        c.argument('admin', options_list=['--admin', '-a'], default=False)
+        c.argument('path', options_list=['--file', '-f'], type=file_type, completer=FilesCompleter(),
+                   default=os.path.join(os.path.expanduser('~'), '.kube', 'config'))
+
 
 def _get_default_install_location(exe_name):
     system = platform.system()
@@ -126,7 +142,7 @@ def _get_default_install_location(exe_name):
         if not home_dir:
             return None
         install_location = os.path.join(home_dir, r'.azure-{0}\{0}.exe'.format(exe_name))
-    elif system == 'Linux' or system == 'Darwin':
+    elif system in ('Linux', 'Darwin'):
         install_location = '/usr/local/bin/{}'.format(exe_name)
     else:
         install_location = None